requirements.txt
python-dotenv==1.0.0
groq==0.4.2
streamlit==1.32.0 